Switch from Process.fork to Process.spawn
CoW didn't save much but caused forking errors.

+++ spec/specwrk_spec.rb
@@ -37,24 +37,22 @@ RSpec.describe Specwrk do
     end
   end
 
-  if ENV.fetch("SPECWRK_SRV_URI", "").include?("localhost") && ENV["SPECWRK_FORKED"]
-    describe "a test that only passes on the second retry on the same instance (assume max retries > 0)" do
-      it "should succeed on the 2nd run" do
-        file = File.join(Dir.tmpdir, "specwrk.retry")
+  describe "a test that only passes on the second retry on the same instance (assume max retries > 0)" do
+    it "should succeed on the 2nd run" do
+      file = File.join(Dir.tmpdir, "specwrk.retry")
 
-        count = if File.exist?(file)
-          JSON.parse(File.read(file))
-        else
-          0
-        end
+      count = if File.exist?(file)
+        JSON.parse(File.read(file))
+      else
+        0
+      end
 
-        File.write(file, JSON.generate(count + 1))
-        if count.zero?
-          expect(true).to eq(false)
-        else
-          expect(true).to eq(true)
-          FileUtils.rm_rf(file)
-        end
+      File.write(file, JSON.generate(count + 1))
+      if count.zero? && ENV.fetch("SPECWRK_SRV_URI", "").include?("localhost") && ENV["SPECWRK_FORKED"]
+        expect(true).to eq(false)
+      else
+        expect(true).to eq(true)
+        FileUtils.rm_rf(file)
       end
     end
   end
